refactor: give passive Home takeover text its own presentation owner
Green pin BEFORE move: f7e6da37d13a19d1da7c3d387870b2e44b2f2499,
test/hero_takeover_info_origin_test.dart, FIRST four public cases PASS.
Tested raw357ddefa80fec8ddd951e26cfe1658f9c0d9e4a8cc16336e07b247db810a2ac8;
Git LF5c6fbb71f4d4bbf772b32b166844b9018c8aa4a901fb74472f39e66d7bd5ae34.
Normalized raw==Gitblob; no origin edit. Original3cc/static9RED preserved:
one realtheme constructor+8references corrected before first behavior run.

Exact accepted product patch7aad5b794cfaa71b01bfcd4979f237393afbc33264b5dbbbf09590e947f293d8
and separate inventory7e3cc2be2743e92a16f2ae6f7313d8174b6f017009d28eae880d66e01973abd7.
Only host, new hero_takeover_info.dart and existing shape_manifest_test.dart.

All origin body differences:
- Host191line _buildTakeoverInfoOverlay becomes HeroTakeoverInfo.build(context)
  with @override and local themeContext=context. Three host notifier aliases
  become borrowed final ValueListenables heroItem/heroEnriched/takeoverSignal.
- Helper calls pass themeContext to renamed _metaText/_metaDot. Their12/14
  line bodies are EXACT; only names/signatures add context parameter. Overlay
  reverses exactly after those aliases/signature/helper-call substitutions.
- Helpers use OWNER OUTER BUILD context, not shadowing nested VLB context;
  each still reads AppThemeScope at metadata assembly. Outer app lookup stays
  before item->enriched VLBs. No blanket palette snapshot. Intentional host
  context->owner dependency and method->widget build phase change explicit;
  finite theme-only then metadata pin, not arbitrary override equivalence.
- Sole host conditional stays _heroTrailerActive/same Stack slot; inline
  new owner receives three original stable refs. Standard optional key is a
  NEW widget API, not alteration of existing constructor. No callbacks,
  host-context parameter, private State, .value snapshot or resource owner.
- Attached origin comments preserved; new owner imports Flutter material,
  foundation, GoogleFonts, StremioMeta and AppThemeScope. One host import,
  no adjacent cleanup. Recursive actual import graph99/no host backedge.
- Native layers, recede/veil, bridge/keyboard/focus/engine lifetimes unchanged.
  Null/item/enriched/takeover nesting, metadata fallback/year/zero-rating/
  first3genres, capturedText reuse, .001 threshold, every ease/segment/rise,
  literal/color/shape and maxWidth720 remain unchanged.

Inventory: hostzero retained, newownerzero added to SAME swept map, exact
new membership/token test. One app.shape.br(4) moves host4->3/new1; total490,
floor490 unchanged, zero radius allowance growth, all other loops unchanged.

Verification on pinned Flutter3.44.8/native Dart sibling, --no-pub:
C:/Users/hunth/sdks/flutter-3.44.8/flutter/bin/flutter.bat test --no-pub test/hero_takeover_info_origin_test.dart test/hero_spotlight_origin_test.dart test/hero_presenter_origin_test.dart test/hero_presenter_runtime_test.dart test/hero_trailer_backdrop_test.dart test/theme/shape_manifest_test.dart --reporter json
ONE six-suite batch105PASS1knownFAIL0skip, exit1/donefalse,10.345s/no180timeout;
widgetdefault600 unchanged. Sole exact BASELINE_ALLOWLIST line7 pair:
test/theme/shape_manifest_test.dart :: no swept file grows a new bare radius: lib/widgets/tv_sidebar_nav.dart
Raw4329d82471da7e562eacebde4832b5cefd35ecf7a81c50b42c0b0ab3fc0216fe.
This is not an all-green batch and no allowance was changed.

Exactly three temporary actual-file probes, each same Flutter command scoped
to test/theme/shape_manifest_test.dart --plain-name <named case> --reporter json:
1 append valid top-level BorderRadius.circular(4) to NEW owner; its existing
  per-file parser case failed intended0->1; raw6cb52e358560a4694b15e47760858297303eab22d9f5dd6ae0d311af27658c2e.
2 same actual valid literal in HOST; host parser case failed intended0->1;
  rawbd753cc5fee5e69bfc0d2704c5b904286d5e1d5c7ec9518ba6610d024f3cab32.
3 remove actual newowner mapentry; Home takeover owner remains swept with its
  moved token failed intendednullvs0; rawa82b445faacbc6e5f87ba99f5f39509cae7d4548c4f3d48ff6a8a2db595d27b1.
Every file restored in finally and verified rawhash before next probe.
ONE pristine3-name confirmation PASS, rawa5236e6fc0708fc99f32df969a5c409878f7e8e166e6870318bedf36a16f5754.

C:/Users/hunth/sdks/flutter-3.44.8/flutter/bin/flutter.bat analyze --no-pub lib/screens/search_screen.dart lib/screens/search/hero_takeover_info.dart test/hero_takeover_info_origin_test.dart test/theme/shape_manifest_test.dart
Four inheritedcacheExtent infos, zero new; rawb2e03f0fad3d7b47c34b0b840d90827750b56fb493a6c5d250c03c5308a4e103.
With pinned native Dart on process PATH: py -3 tool/analyze_baseline.py
431/449 zeroNew (0error76warning355info), exit0, unchanged baseline;
raw5c6c4468fec495c465d743e640412482e0890c5824a5bdcf3a69f6d283f6ddf5.
C:/Users/hunth/sdks/flutter-3.44.8/flutter/bin/cache/dart-sdk/bin/dart.exe tool/check_layering.dart --json
Same checker on originf7e6 archivedlib via --root .dart_tool/takeover-origin-prep/product/parent-source --ceiling-file tool/layering_baseline.txt --json
Both56/77/exit0 and rawidentical e192d6f4153f156865f1470c36ef6999849c5f673cb7e2ac3fac0890e5deebc5.

Ignored application scope verifier initially used only git diff --name-only,
omitting expected untracked newowner, and asserted after exactpatches applied.
Read-only tracked+untracked union/sourcehash proof resolved evidence: no drift,
reapplication or sourcefix. Preserved note. Origin raw-vs-LF verifier note also
preserved. C0 independently accepted exact author raw/probes/source restoration;
independent production runtime remains separate, not claimed here.

Did we make a difference?
Three host methods217lines now have one passive presentation owner/3borrowed
refs/0callbacks. Removed block229, replacement+import6 =>host-223; newowner252,
wholeproduction+29. Inventory adds7 separately. No deletion-only claim.
ZERO seven-stage credit: this is classic Home, does not reduce remaining353.

Is there more we could do?
Native/shell authority, frame/row/focus, recede/veil remain with current owners;
not reopened or waived. No new forwarder; hostconditional/borrowedrefs remain
intentional composition. StrictSearch remains OPEN. Finite signal-driven UI
proof is not native playback/engine-start/transport/alltheme/allinterleaving
proof. Later226/main union must preserve Atrium; no publication/merge here.
